+++ src/str_mut_signatures/__init__.py
@@ -34,7 +34,7 @@ from .nmf.plot import (
 
 __author__ = "Olesia Kondrateva"
 __email__ = "xkdnoa@gmail.com"
-__version__ = "2.0.0"
+__version__ = "2.0.1"
 
 __all__ = [
     # VCF → mutations